Create Default file browser plugin with a YDrive
