+++ packages/rtc-extension/package.json
@@ -57,10 +57,13 @@
     "@jupyterlab/apputils": "^4.0.0-alpha.18",
     "@jupyterlab/collaboration": "^4.0.0-alpha.18",
     "@jupyterlab/coreutils": "^6.0.0-alpha.18",
+    "@jupyterlab/docprovider": "^4.0.0-alpha.18",
+    "@jupyterlab/filebrowser": "^4.0.0-alpha.18",
     "@jupyterlab/services": "^7.0.0-alpha.18",
     "@jupyterlab/statedb": "^4.0.0-alpha.18",
     "@jupyterlab/translation": "^4.0.0-alpha.18",
     "@jupyterlab/ui-components": "^4.0.0-alpha.33",
+    "@lumino/commands": "^2.0.0-alpha.6",
     "@lumino/widgets": "^2.0.0-alpha.6",
     "y-protocols": "^1.0.5",
     "y-websocket": "^1.3.15",
@@ -80,6 +83,19 @@
   "jupyterlab": {
     "extension": true,
     "schemaDir": "./schema",
-    "outputDir": "../../jupyterlab_rtc/labextension"
+    "outputDir": "../../jupyterlab_rtc/labextension",
+    "disabledExtensions": [
+      "@jupyterlab/filebrowser-extension:defaultFileBrowser"
+    ],
+    "sharedPackages": {
+      "@jupyterlab/collaboration": {
+        "bundled": true,
+        "singleton": true
+      },
+      "@jupyterlab/docprovider": {
+        "bundled": true,
+        "singleton": true
+      }
+    }
   }
 }
